+++ src/main/java/frc/robot/automations/shootingIntakeSlammer/HopperTrackerIO.java
@@ -1,0 +1,16 @@
+package frc.robot.automations.shootingIntakeSlammer;
+
+import org.littletonrobotics.junction.AutoLog;
+
+public interface HopperTrackerIO {
+	@AutoLog
+	public static class HopperTrackerIOInputs {
+		public int ballCount;
+		public boolean intakeDeployed;
+	}
+
+	public default void updateInputs(HopperTrackerIOInputs inputs) {}
+
+	default void setBallCount(int balls) {}
+	default void setIntakeDeployed(boolean deployed) {}
+}
